refactor: split SB3 training helper boundaries
