--- .github/workflows/refactor-sb3-training-boundaries.yml
@@ -1,123 +1,0 @@
-name: Refactor SB3 training boundaries
-
-on:
-  push:
-    branches:
-      - agent/split-sb3-training-boundaries
-
-permissions:
-  contents: write
-
-concurrency:
-  group: refactor-sb3-training-boundaries
-  cancel-in-progress: true
-
-jobs:
-  refactor:
-    runs-on: ubuntu-latest
-    timeout-minutes: 90
-    steps:
-      - name: Checkout exact branch head
-        uses: actions/checkout@34e114876b0b11c390a56381ad16ebd13914f8d5
-        with:
-          ref: ${{ github.sha }}
-          fetch-depth: 0
-          persist-credentials: true
-
-      - name: Apply deterministic boundary extraction
-        run: python .github/refactor_sb3_training_boundaries.py
-
-      - name: Set up uv
-        uses: astral-sh/setup-uv@d4b2f3b6ecc6e67c4457f6d3e41ec42d3d0fcb86
-        with:
-          python-version: "3.12"
-          enable-cache: true
-
-      - name: Install verification dependencies
-        run: uv sync --extra dev --extra train-sb3
-
-      - name: Format and lint touched files
-        shell: bash
-        run: |
-          set -euo pipefail
-          files=(
-            trade_rl/integrations/sb3_training.py
-            trade_rl/integrations/sb3_runtime.py
-            trade_rl/integrations/sb3_environment.py
-            trade_rl/integrations/sb3_behavior_cloning.py
-            trade_rl/integrations/sb3_teacher_pipeline.py
-            tests/architecture/test_sb3_training_boundaries.py
-            tests/architecture/test_ownership_boundaries.py
-            tests/integrations/test_sb3_training.py
-            tests/learning/test_episode_teacher_integration.py
-          )
-          uv run ruff check --fix "${files[@]}"
-          uv run ruff format "${files[@]}"
-
-      - name: Run focused boundary regression suite
-        shell: bash
-        run: |
-          set -euo pipefail
-          uv run pytest \
-            tests/architecture/test_sb3_training_boundaries.py \
-            tests/architecture/test_ownership_boundaries.py \
-            tests/integrations/test_sequence_runtime_acceleration.py \
-            tests/integrations/test_action_head_bc_routing.py \
-            tests/integrations/test_parallel_sequence_subprocess_smoke.py \
-            tests/integrations/test_sb3_training.py \
-            tests/integrations/test_sb3_training_transfer.py \
-            tests/integrations/test_sb3_lagrangian_probe_backend.py \
-            tests/integrations/test_sb3_cost_critic_backend.py \
-            tests/integrations/test_sb3_lagrangian_backend.py \
-            tests/integrations/test_sb3_constraint_info.py \
-            tests/learning/test_episode_teacher_integration.py \
-            -q
-
-      - name: Run focused static and architecture gates
-        shell: bash
-        run: |
-          set -euo pipefail
-          uv run mypy \
-            trade_rl/integrations/sb3_training.py \
-            trade_rl/integrations/sb3_runtime.py \
-            trade_rl/integrations/sb3_environment.py \
-            trade_rl/integrations/sb3_behavior_cloning.py \
-            trade_rl/integrations/sb3_teacher_pipeline.py
-          uv run lint-imports
-          uv run vulture trade_rl --min-confidence 100
-          test "$(wc -c < trade_rl/integrations/sb3_training.py)" -lt 61440
-
-      - name: Remove temporary automation and verify scope
-        shell: bash
-        run: |
-          set -euo pipefail
-          git rm \
-            .github/workflows/refactor-sb3-training-boundaries.yml \
-            .github/refactor_sb3_training_boundaries.py
-          forbidden=(
-            docs/implementation/2026-08-11-universal-instrument-artifact-materialization-plan.md
-            tests/workflows/test_postgres_universal_instrument_artifacts.py
-            tests/workflows/test_universal_instrument_artifacts.py
-            trade_rl/workflows/universal_instrument_artifacts.py
-          )
-          changed="$(git diff --name-only HEAD)"
-          for path in "${forbidden[@]}"; do
-            if grep -Fxq "$path" <<<"$changed"; then
-              echo "unexpected overlap with PR #385: $path" >&2
-              exit 1
-            fi
-          done
-          git diff --check
-
-      - name: Commit and push verified refactor
-        shell: bash
-        env:
-          TARGET_BRANCH: agent/split-sb3-training-boundaries
-        run: |
-          set -euo pipefail
-          git config user.name "Shuntatsu"
-          git config user.email "148511409+shuntatsu@users.noreply.github.com"
-          git add -A
-          git diff --cached --check
-          git commit -m "refactor: split SB3 training helper boundaries"
-          git push origin "HEAD:${TARGET_BRANCH}"
